implemented existing control to display transparency correctly
