--- src/components/site/EmphasizedSectionTransparency.js
@@ -1,8 +1,0 @@
-import React from "react";
-
-const EmphasizedSectionTransparency = props => (
-  <div className="container">
-   
-  </div>
-);
-export default EmphasizedSectionTransparency;
