--- src/components/site/EmphasizedSectionTransparency.md
@@ -1,4 +1,0 @@
-```jsx
-import SectionTransparency from "./SectionTransparency";
-  <SectionTransparency></SectionTransparency>
-```
