+++ src/components/site/SectionTransparency.js
@@ -1,11 +1,26 @@
 import React from 'react';
+import PropTypes from 'prop-types';
+import IconHeading from '../basic/IconHeading';
 
 const SectionTransparency = (props) => (
-	<div className="align-text-center">
-		<div className="col-md-12 mb-5 testing-style-drk">
-			<i className="icon-dark icon-section-header-light fa fa-search-plus" aria-hidden="true" />
+	<div>
+		<div>
+			<IconHeading text="TRANSPARENCY" icon="fa fa-search-plus" />
 		</div>
-		<h3 className="dg_section_transparency">Transparency</h3>
+		<div>{props.children}</div>
 	</div>
 );
+SectionTransparency.propTypes = {
+	/**  A text description of the action the icon provides. */
+	description: PropTypes.string.isRequired,
+	/** Html href property */
+	href: PropTypes.string.isRequired,
+	/** Font Awesome icon class */
+	icon: PropTypes.string.isRequired,
+	/** icon Text*/
+	text: PropTypes.string,
+	/** Visual type of the icon link, possible values are `rectangle` and `circle` */
+	type: PropTypes.string
+};
+
 export default SectionTransparency;
